requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
pymupdf
Pillow
accelerate
torch
